[Fix-5519] The executor api, projectName change to projectCode (#5863)
* fix: the executor api, projectName change to projectCode

* fix checkstyle

* fix: when process definition is null, return error message
